requirements.txt
fastapi
uvicorn
groq
python-dotenv
requests
apscheduler
chromadb
beautifulsoup4